Renamed files to match lower-kebab-case and removed unnecessary files.
